requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
dnspython==1.16.0
pymongo==3.11.0
pytz==2020.1
sqlparse==0.2.4

